+++ src/components/List/list.css
@@ -1,17 +1,15 @@
-.pagination-button {
-  padding: 10px;
-  border-radius: 5px;
-  border: 1px solid black;
-  background-color: white;
-}
-.pagination-container {
+
+.details-list-wrapper {
+  min-height: 320px;
   display: flex;
-  flex-direction: row;
-  gap: 10px;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
 }
 
 .details-list {
   list-style-type: none;
+  min-height: 250px;
 
   & .details-item {
     line-height: 1.8;
