+++ src/components/Main/Main.tsx
@@ -1,25 +1,14 @@
 import { List } from '../List';
-import { Spinner } from '../Spinner/Spinner';
 import { Popup } from '../Popup';
 import './main.css';
-import { useSearchParams } from 'react-router';
-import { useGetStarWarsPeopleQuery } from '../../redux/api';
+import { Pagination } from '../Pagination';
 //TBD:restructure layoutdue jumping list
 export const Main = () => {
-  const [searchParams] = useSearchParams();
-
-  const { data, isSuccess, isFetching, isError } = useGetStarWarsPeopleQuery(
-    searchParams.toString()
-  );
-
   return (
     <div className="main-container">
       <header className="main-header">Results</header>
-      {isFetching && <Spinner />}
-      {isError && <span className="errorText">Something were wrong ...</span>}
-      {!isFetching && !isError && isSuccess && (
-        <List people={data.results} previous={data.previous} next={data.next} />
-      )}
+      <List />
+      <Pagination />
       <Popup />
     </div>
   );
